+++ menu_utils.h
@@ -1,0 +1,6 @@
+#ifndef MENU_UTILS_H_INCLUDED
+#define MENU_UTILS_H_INCLUDED
+
+int inlineOneButtonSelect(int button_width, char *buttons[], int numButtons, int left_spacing, int button_spacing, int is_3_line, int top_offset);
+
+#endif /* MENU_UTILS_H_INCLUDED */
